refactor: split shop validators
