--- apps/cms/src/services/shops/validation.ts
@@ -1,269 +1,0 @@
-import { z } from "zod";
-import { localeSchema, type Locale, aiCatalogFieldSchema } from "@acme/types";
-import { shopSchema, type ShopForm } from "../../actions/schemas";
-import {
-  parseFilterMappings,
-  parsePriceOverrides,
-  parseLocaleOverrides,
-} from "./formData";
-import { formDataEntries, formDataToObject } from "../../utils/formData";
-
-export type { ShopForm };
-
-export function parseShopForm(formData: FormData): {
-  data?: ShopForm;
-  errors?: Record<string, string[]>;
-} {
-  const themeDefaultsRaw = formData.get("themeDefaults") as string | null;
-  const themeOverridesRaw = formData.get("themeOverrides") as string | null;
-
-  const entries = Array.from(formDataEntries(formData)).filter(
-    ([k]) =>
-      ![
-        "filterMappingsKey",
-        "filterMappingsValue",
-        "priceOverridesKey",
-        "priceOverridesValue",
-        "localeOverridesKey",
-        "localeOverridesValue",
-      ].includes(k)
-  ) as [string, FormDataEntryValue][];
-
-  const parsed = shopSchema.safeParse({
-    ...Object.fromEntries(entries),
-    themeDefaults: themeDefaultsRaw ?? "{}",
-    themeOverrides: themeOverridesRaw ?? "{}",
-    trackingProviders: formData.getAll("trackingProviders"),
-    filterMappings: parseFilterMappings(formData),
-    priceOverrides: parsePriceOverrides(formData),
-    localeOverrides: parseLocaleOverrides(formData),
-  });
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-const seoSchema = z
-  .object({
-    locale: localeSchema,
-    title: z.string().min(1, "Required"),
-    description: z.string().optional().default(""),
-    image: z.string().url().optional(),
-    alt: z.string().optional(),
-    canonicalBase: z.string().url().optional(),
-    ogUrl: z.string().url().optional(),
-    twitterCard: z
-      .enum(["summary", "summary_large_image", "app", "player"])
-      .optional(),
-  })
-  .strict();
-
-export function parseSeoForm(formData: FormData): {
-  data?: z.infer<typeof seoSchema>;
-  errors?: Record<string, string[]>;
-} {
-  const parsed = seoSchema.safeParse(formDataToObject(formData));
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-const generateSchema = z
-  .object({
-    id: z.string().min(1),
-    locale: localeSchema,
-    title: z.string().min(1),
-    description: z.string().min(1),
-  })
-  .strict();
-
-export function parseGenerateSeoForm(formData: FormData): {
-  data?: z.infer<typeof generateSchema>;
-  errors?: Record<string, string[]>;
-} {
-  const parsed = generateSchema.safeParse(formDataToObject(formData));
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-const currencyTaxSchema = z
-  .object({
-    currency: z.string().length(3, "Required"),
-    taxRegion: z.string().min(1, "Required"),
-  })
-  .strict();
-
-export function parseCurrencyTaxForm(formData: FormData): {
-  data?: z.infer<typeof currencyTaxSchema>;
-  errors?: Record<string, string[]>;
-} {
-  const parsed = currencyTaxSchema.safeParse(formDataToObject(formData));
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-const depositSchema = z
-  .object({
-    enabled: z.preprocess((v) => v === "on", z.boolean()),
-    intervalMinutes: z.coerce.number().int().min(1, "Must be at least 1"),
-  })
-  .strict();
-
-export function parseDepositForm(formData: FormData): {
-  data?: z.infer<typeof depositSchema>;
-  errors?: Record<string, string[]>;
-} {
-  const parsed = depositSchema.safeParse(formDataToObject(formData));
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-const reverseLogisticsSchema = z
-  .object({
-    enabled: z.preprocess((v) => v === "on", z.boolean()),
-    intervalMinutes: z.coerce.number().int().min(1, "Must be at least 1"),
-  })
-  .strict();
-
-export function parseReverseLogisticsForm(formData: FormData): {
-  data?: z.infer<typeof reverseLogisticsSchema>;
-  errors?: Record<string, string[]>;
-} {
-  const parsed = reverseLogisticsSchema.safeParse(formDataToObject(formData));
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-const returnsSchema = z
-  .object({
-    enabled: z.preprocess((v) => v === "on", z.boolean()),
-    bagEnabled: z
-      .preprocess((v) => v === "on", z.boolean())
-      .optional(),
-    homePickupEnabled: z
-      .preprocess((v) => v === "on", z.boolean())
-      .optional(),
-  })
-  .strict();
-
-export function parseUpsReturnsForm(formData: FormData): {
-  data?: z.infer<typeof returnsSchema>;
-  errors?: Record<string, string[]>;
-} {
-  const parsed = returnsSchema.safeParse(formDataToObject(formData));
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-const premierDeliverySchema = z
-  .object({
-    regions: z.array(z.string().min(1)).default([]),
-    windows: z.array(z.string().regex(/^\d{2}-\d{2}$/)).default([]),
-    carriers: z.array(z.string().min(1)).default([]),
-    surcharge: z.coerce.number().int().nonnegative().optional(),
-    serviceLabel: z.string().optional(),
-  })
-  .strict();
-
-export function parsePremierDeliveryForm(formData: FormData): {
-  data?: z.infer<typeof premierDeliverySchema>;
-  errors?: Record<string, string[]>;
-} {
-  const data = {
-    regions: formData
-      .getAll("regions")
-      .map(String)
-      .map((v) => v.trim())
-      .filter(Boolean),
-    windows: formData
-      .getAll("windows")
-      .map(String)
-      .map((v) => v.trim())
-      .filter(Boolean),
-    carriers: formData
-      .getAll("carriers")
-      .map(String)
-      .map((v) => v.trim())
-      .filter(Boolean),
-    surcharge: formData.get("surcharge"),
-    serviceLabel: formData.get("serviceLabel")?.toString().trim() || undefined,
-  };
-  const parsed = premierDeliverySchema.safeParse(data);
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-const aiCatalogFormSchema = z
-  .object({
-    enabled: z.preprocess((v) => v === "on", z.boolean()),
-    pageSize: z.coerce.number().int().positive(),
-    fields: z.array(aiCatalogFieldSchema),
-  })
-  .strict();
-
-export function parseAiCatalogForm(formData: FormData): {
-  data?: z.infer<typeof aiCatalogFormSchema>;
-  errors?: Record<string, string[]>;
-} {
-  const data = {
-    enabled: formData.get("enabled"),
-    pageSize: formData.get("pageSize"),
-    fields: formData.getAll("fields"),
-  };
-  const parsed = aiCatalogFormSchema.safeParse(data);
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-const stockAlertFormSchema = z
-  .object({
-    recipients: z.array(z.string().email()),
-    webhook: z.string().url().optional(),
-    threshold: z.coerce.number().int().min(1, "Must be at least 1").optional(),
-  })
-  .strict();
-
-export function parseStockAlertForm(formData: FormData): {
-  data?: z.infer<typeof stockAlertFormSchema>;
-  errors?: Record<string, string[]>;
-} {
-  const data = {
-    recipients: String(formData.get("recipients") ?? "")
-      .split(",")
-      .map((r) => r.trim())
-      .filter(Boolean),
-    webhook: formData.get("webhook")?.toString().trim() || undefined,
-    threshold: formData.get("threshold"),
-  };
-  const parsed = stockAlertFormSchema.safeParse(data);
-  if (!parsed.success) {
-    return { errors: parsed.error.flatten().fieldErrors };
-  }
-  return { data: parsed.data };
-}
-
-export type SeoForm = z.infer<typeof seoSchema> & { locale: Locale; };
-export type GenerateSeoForm = z.infer<typeof generateSchema>;
-export type CurrencyTaxForm = z.infer<typeof currencyTaxSchema>;
-export type DepositForm = z.infer<typeof depositSchema>;
-export type ReverseLogisticsForm = z.infer<typeof reverseLogisticsSchema>;
-export type UpsReturnsForm = z.infer<typeof returnsSchema>;
-export type PremierDeliveryForm = z.infer<typeof premierDeliverySchema>;
-export type AiCatalogForm = z.infer<typeof aiCatalogFormSchema>;
-export type StockAlertForm = z.infer<typeof stockAlertFormSchema>;
