+++ apps/cms/src/services/shops/validation/parseCurrencyTaxForm.ts
@@ -1,0 +1,22 @@
+import { z } from "zod";
+import { formDataToObject } from "../../../utils/formData";
+
+const currencyTaxSchema = z
+  .object({
+    currency: z.string().length(3, "Required"),
+    taxRegion: z.string().min(1, "Required"),
+  })
+  .strict();
+
+export function parseCurrencyTaxForm(formData: FormData): {
+  data?: z.infer<typeof currencyTaxSchema>;
+  errors?: Record<string, string[]>;
+} {
+  const parsed = currencyTaxSchema.safeParse(formDataToObject(formData));
+  if (!parsed.success) {
+    return { errors: parsed.error.flatten().fieldErrors };
+  }
+  return { data: parsed.data };
+}
+
+export type CurrencyTaxForm = z.infer<typeof currencyTaxSchema>;
