+++ apps/cms/src/services/shops/validation/parseUpsReturnsForm.ts
@@ -1,0 +1,27 @@
+import { z } from "zod";
+import { formDataToObject } from "../../../utils/formData";
+
+const returnsSchema = z
+  .object({
+    enabled: z.preprocess((v) => v === "on", z.boolean()),
+    bagEnabled: z
+      .preprocess((v) => v === "on", z.boolean())
+      .optional(),
+    homePickupEnabled: z
+      .preprocess((v) => v === "on", z.boolean())
+      .optional(),
+  })
+  .strict();
+
+export function parseUpsReturnsForm(formData: FormData): {
+  data?: z.infer<typeof returnsSchema>;
+  errors?: Record<string, string[]>;
+} {
+  const parsed = returnsSchema.safeParse(formDataToObject(formData));
+  if (!parsed.success) {
+    return { errors: parsed.error.flatten().fieldErrors };
+  }
+  return { data: parsed.data };
+}
+
+export type UpsReturnsForm = z.infer<typeof returnsSchema>;
